Minor modifications to the daily check-in template
The idea is to try and build a filter of templates for a fast summary
at the end of the day. I will filter for all entries with the tags
:instant: and the status not= DONE / MEETING, and will auto copy their
contents into an end-of-day note which will be the :summary: of the
day.

How to do this? I will probably write a YASnippet template for this,
since this is what I've done for templates which are large / execute
complex elisp.

One task that I should look into later: Can `org-mode capture`
templates support complex `elisp` execution? If so, I might be able to
get rid of YASnippet altogether in favor of capture. This is not high
on my priority list at the moment though.